[#45] NicknameSettingView에서 성별, 나이 버튼 제거
