+++ EUMMEYO/View/Nickname/NicknameSettingView.swift
@@ -21,22 +21,6 @@ struct NicknameSettingView: View {
     private var isNicknameValid: Bool {
         !nickname.trimmingCharacters(in: .whitespaces).isEmpty
     }
-
-    
-    // 생년월일 날짜 포맷터
-    private var dateFormatter: DateFormatter {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy. MM. dd"
-        return formatter
-    }
-    
-    // 전송용 생년월일 날짜 포맷터
-    private var isoDateFormatter: ISO8601DateFormatter {
-        let formatter = ISO8601DateFormatter()
-        formatter.timeZone = TimeZone(identifier: "Asia/Seoul") // KST 설정
-        return formatter
-    }
-
     
     var body: some View {
         VStack(spacing: 10) {
@@ -70,39 +54,7 @@ struct NicknameSettingView: View {
             
             Spacer()
                 .frame(height: 10)
-            
-            Text("생년월일")
-                .frame(maxWidth: .infinity, alignment: .leading)
-            
-            Button {
-                isDatePickerActive.toggle()
-            } label: {
-                Text("\(dateFormatter.string(from: birthdate))")
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .foregroundColor(Color.mainBlack)
-                    .padding()
-                    .frame(maxWidth: .infinity)
-                    .background(Color.white)
-                    .overlay(
-                        RoundedRectangle(cornerRadius: 10)
-                            .stroke(lineWidth: 1.5)
-                            .foregroundColor(.black)
-                    )
-            }
-            
-            Spacer()
-                .frame(height: 10)
-            
-            Text("성별")
-                .frame(maxWidth: .infinity, alignment: .leading)
-            
-            HStack {
-                GenderButton(gender: .male, isSelected: $selectedGender)
-                GenderButton(gender: .female, isSelected: $selectedGender)
-                GenderButton(gender: .other, isSelected: $selectedGender)
-            }
-            
-            
+        
             Spacer()
             
             Button {
@@ -115,13 +67,8 @@ struct NicknameSettingView: View {
                         } else {
                             // 업데이트 성공
                             nicknameMessage = ""
-                            //authViewModel.send(action: .updateUserNickname(nickname))
-                            
-                            let birthdayString = isoDateFormatter.string(from: birthdate)
-                            print("디버깅: \(birthdayString)")
-                            let genderString = selectedGender.rawValue
-                            authViewModel.send(action: .updateUserInfo(nickname, birthdayString, genderString))
-
+                        
+                            authViewModel.send(action: .updateUserInfo(nickname, "", ""))
                         }
                     })
                 }
@@ -140,91 +87,11 @@ struct NicknameSettingView: View {
         }.padding(.horizontal, 25)
         .offset(x: slideOffset) // x축 오프셋 적용
         .onAppear {
-//            withAnimation(.easeOut(duration: 0.4)) { // 0.3초 동안 easeOut 애니메이션
-                slideOffset = 0 // 오프셋을 0으로 만들어 화면 중앙으로 이동
-//            }
-        }
-        .sheet(isPresented: $isDatePickerActive) {
-            BirthdayPickerView(birthdate: $birthdate)
-                .presentationDetents([.fraction(0.5)])
+            slideOffset = 0 // 오프셋을 0으로 만들어 화면 중앙으로 이동
         }
     }
 }
 
 #Preview {
     NicknameSettingView()
-}
-
-
-// MARK: - 생년월일 선택
-private struct BirthdayPickerView: View {
-    @Binding var birthdate: Date
-    @Environment(\.dismiss) private var dismiss // sheet 닫기
-    
-    fileprivate var body: some View {
-        VStack {
-             Text("생년월일 선택")
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .padding(.leading, 50)
-                .font(.system(size: 20, weight: .bold))
-                
-            DatePicker(
-                "생년얼일",
-                selection: $birthdate,
-                in: ...Date(), // 현재 날짜까지 선택 가능
-                displayedComponents: .date
-            )
-            .datePickerStyle(.wheel)
-            .labelsHidden()
-            .padding()
-            .overlay(
-                RoundedRectangle(cornerRadius: 20)
-                    .stroke(.black, lineWidth: 2)
-            )
-            .environment(\.locale, Locale(identifier: "ko_KR"))
-            
-            Spacer()
-                .frame(height: 30)
-            
-            Button {
-                dismiss()
-            } label: {
-                Text("완료")
-                    .padding()
-                    .frame(maxWidth: .infinity)
-                    .foregroundColor(Color.white)
-                    .background(Color.mainBlack)
-                    .cornerRadius(20)
-            }
-            .padding(.horizontal, 40)
-            
-        }
-    }
-}
-
-// MARK: - 성별 버튼
-private struct GenderButton: View {
-    
-    // 성별
-    var gender: Gender
-    
-    // 선택유무
-    @Binding var isSelected: Gender
-    
-    fileprivate var  body: some View {
-        Button {
-            isSelected = gender
-        } label: {
-            Text(gender.rawValue)
-                .foregroundColor(isSelected == gender ? Color.white : Color.black)
-                .padding()
-                .frame(maxWidth: .infinity)
-                .overlay {
-                    RoundedRectangle(cornerRadius: 10)
-                        .stroke(.black, lineWidth: 2)
-                }
-                .background(isSelected == gender ? Color.mainBlack : Color.white) // 선택 여부에 따라 배경색 변경
-                .clipShape(RoundedRectangle(cornerRadius: 10))
-        }
-    }
 }
